requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
pydantic[email]
langchain-core
langchain-groq
langgraph
tavily-python
feedparser
requests
beautifulsoup4
jinja2
resend
apscheduler
psycopg2-binary